# internal/app/swagger/swagger.json
{"swagger":"2.0","info":{"title":"GinAdmin","description":"RBAC scaffolding based on GIN + GORM + CASBIN.","version":"3.1.1","contact":{"name":"LyricTian","email":"tiannianshou@gmail.com"},"license":{}},"host":"127.0.0.1:10088","basePath":"/","schemes":["http","https"],"consumes":["application/json"],"produces":["application/json"],"paths":{"/api/v1/demos":{"post":{"tags":["Demo"],"summary":"创建数据","operationId":"Demo.Create","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}},"400":{"description":"{error:{code:0,message:无效的请求参数}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/api/v1/demos/{id}":{"get":{"tags":["Demo"],"summary":"查询指定数据","operationId":"Demo.Get","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"404":{"description":"{error:{code:0,message:资源不存在}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}},"put":{"tags":["Demo"],"summary":"更新数据","operationId":"Demo.Update","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}},"400":{"description":"{error:{code:0,message:无效的请求参数}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}},"delete":{"tags":["Demo"],"summary":"删除数据","operationId":"Demo.Delete","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"}],"responses":{"200":{"description":"{status:OK}","schema":{"$ref":"#/definitions/HTTPStatus","title":"HTTPStatus","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/api/v1/demos/{id}/disable":{"patch":{"tags":["Demo"],"summary":"禁用数据","operationId":"Demo.Disable","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"}],"responses":{"200":{"description":"{status:OK}","schema":{"$ref":"#/definitions/HTTPStatus","title":"HTTPStatus","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/api/v1/demos/{id}/enable":{"patch":{"tags":["Demo"],"summary":"启用数据","operationId":"Demo.Enable","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"path","name":"id","description":"记录ID","required":true,"type":"string"}],"responses":{"200":{"description":"{status:OK}","schema":{"$ref":"#/definitions/HTTPStatus","title":"HTTPStatus","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}},"/api/v1/demos?q=page":{"get":{"tags":["Demo"],"summary":"查询分页数据","operationId":"Demo.QueryPage","parameters":[{"in":"header","name":"Authorization","description":"Bearer 用户令牌","type":"string"},{"in":"query","name":"current","description":"分页索引","required":true,"type":"integer","format":"int32","default":1},{"in":"query","name":"pageSize","description":"分页大小","required":true,"type":"integer","format":"int32","default":10},{"in":"query","name":"code","description":"编号","type":"string"},{"in":"query","name":"name","description":"名称","type":"string"},{"in":"query","name":"status","description":"状态(1:启用 2:停用)","type":"integer","format":"int32"}],"responses":{"200":{"description":"查询结果：{list:列表数据,pagination:{current:页索引,pageSize:页大小,total:总数量}}","schema":{"type":"array","items":{"$ref":"#/definitions/Demo","title":"Demo","type":"object"}}},"400":{"description":"{error:{code:0,message:未知的查询类型}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"401":{"description":"{error:{code:0,message:未授权}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}},"500":{"description":"{error:{code:0,message:服务器错误}}","schema":{"$ref":"#/definitions/HTTPError","title":"HTTPError","type":"object"}}}}}},"definitions":{"Demo":{"title":"Demo","required":["code","name","status"],"type":"object","properties":{"code":{"description":"编号","type":"string"},"created_at":{"description":"创建时间","type":"string","format":"date-time"},"creator":{"description":"创建者","type":"string"},"memo":{"description":"备注","type":"string"},"name":{"description":"名称","type":"string"},"record_id":{"description":"记录ID","type":"string"},"status":{"description":"状态(1:启用 2:停用)","type":"integer","format":"int32"}}},"HTTPError":{"title":"HTTPError","required":["error"],"type":"object","properties":{"error":{"$ref":"#/definitions/HTTPErrorItem","description":"错误项","type":"object"}}},"HTTPErrorItem":{"title":"HTTPErrorItem","required":["code","message"],"type":"object","properties":{"code":{"description":"错误码","type":"integer","format":"int32"},"message":{"description":"错误信息","type":"string"}}},"HTTPStatus":{"title":"HTTPStatus","required":["status"],"type":"object","properties":{"status":{"description":"状态(OK)","type":"string"}}}},"tags":[{"name":"Demo","description":"demo"}]}